+++ src/main/java/inventorycount/InventoryCountInteractor.java
@@ -22,12 +22,12 @@ public class InventoryCountInteractor {
 
     }
 
-    public void updateInventoryCount(HashMap<Integer, Integer> newInventoryCount){
+    public void updateInventoryCount(HashMap<Long, Integer> newInventoryCount){
         // pull facility entity
         Facility facility = facilityDb.getFacility(facID);
 
         // update inventory
-        for (int upc : newInventoryCount.keySet()){
+        for (Long upc : newInventoryCount.keySet()){
 
             int newCount = newInventoryCount.get(upc);
 
@@ -44,7 +44,7 @@ public class InventoryCountInteractor {
 
     }
 
-    public HashMap<Integer, Integer> getCurrentCount(){
+    public HashMap<Long, Integer> getCurrentCount(){
         Facility facility = facilityDb.getFacility(facID);
 
         return facility.getInventory();
